requirements.txt
requests==2.31.0
pandas==2.2.0
pytest==7.4.4
hydra-core==1.1.1